Small fixes such that I can pick up work again next time.

+++ src/RoadRegistry/Model/RoadNetworkRevision.cs
@@ -14,12 +14,12 @@ namespace RoadRegistry.Model
             _value = value;
         }
 
-        public int ToInt64() => _value;
+        public int ToInt32() => _value;
         public bool Equals(RoadNetworkRevision other) => _value == other._value;
         public override bool Equals(object other) => other is RoadNetworkRevision revision && Equals(revision);
         public override int GetHashCode() => _value.GetHashCode();
         public static bool operator ==(RoadNetworkRevision left, RoadNetworkRevision right) => left.Equals(right);
         public static bool operator !=(RoadNetworkRevision left, RoadNetworkRevision right) => !left.Equals(right);
-        public static implicit operator long(RoadNetworkRevision instance) => instance._value;
+        public static implicit operator int(RoadNetworkRevision instance) => instance._value;
     }
 }
